#56 feat: Auth 인수 테스트를 위한 Fixture 생성
